Multiple Pass Order Resolution
This is the start of multi-pass order resolution. The issue is visible
in the following example.

Suppose units 1, 2, and 3 are in regions A, B, and C, respectively. The
regions are connected in a chain.
If 1 moves to B, and 2 moves to C, but C remains stationary, all three
should fail. However, due to 3 remaining stationary, 2 fails, which
results in 1 eating 2.

I will have to do multiple passes to ensure actions resolve correctly.
Either that or work out some method of building actions into a
dependency “chain” and then resolving backwards through the chain.